Move table data to another file
